+++ src/views/admin/layouts/baseauth.blade.php
@@ -31,7 +31,7 @@
         </div>
 
         {{-- Start footer scripts --}}
-        {{ HTML::script('packages/jacopo/laravel-authentication-acl/js/jquery-1.10.2.min.js') }}
-        {{ HTML::script('packages/jacopo/laravel-authentication-acl/js/bootstrap.min.js') }}
+        {{ HTML::script('packages/jacopo/laravel-authentication-acl/js/vendor/jquery-1.10.2.min.js') }}
+        {{ HTML::script('packages/jacopo/laravel-authentication-acl/js/vendor/bootstrap.min.js') }}
     </body>
 </html>