+++ src/views/admin/permission/edit.blade.php
@@ -47,7 +47,7 @@ Admin area: edit permission
 @stop
 
 @section('footer_scripts')
-{{HTML::script('packages/jacopo/laravel-authentication-acl/js/slugit.js')}}
+{{HTML::script('packages/jacopo/laravel-authentication-acl/js/vendor/slugit.js')}}
 <script>
     $(".delete").click(function(){
         return confirm("Are you sure to delete this item?");